Add QMP-based HID simulation draft
Add a draft of performing timer-based HID simulation by utilizing
Qemu's QMP interface. An Unix domain socket, which has to be
exposed by the relevant guest, is used to communicate with the
guest and send absolute mouse coordinates. To be able to use
threads the linker option -lthread was added via automake-settings.